Refactor to create CatelogTable.jsx
